Test endpoints fully, migrate float to Numeric and write final documentation to send as the assignment

+++ app/models/expense_model.py
@@ -1,6 +1,7 @@
-from datetime import datetime, timezone
+from datetime import datetime
+from decimal import Decimal
 
-from sqlalchemy import Integer, String, Float, DateTime, ForeignKey
+from sqlalchemy import Integer, String, Numeric, DateTime, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.core.database import Base
@@ -9,7 +10,7 @@ class Expense(Base):
     __tablename__ = "expenses"
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
-    amount: Mapped[float] = mapped_column(Float, nullable=False)
+    amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     description: Mapped[str | None] = mapped_column(String(255), nullable=True)
     date: Mapped[datetime] = mapped_column(
         DateTime,
